requirements.txt
beautifulsoup4==4.11.2
celery==5.2.7
django-bootstrap-v5==1.0.11
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-jquery==3.1.0
django-query==1.0.6
django-redis==5.2.0
django-select2==8.1.1
Django==4.1.7
ezdxf==1.0.2
numpy==1.24.2
pandas==1.5.3
redis==4.5.1
requests==2.28.2
setuptools==60.2.0
sqlparse==0.4.3
XlsxWriter==3.0.9
python-dotenv==1.0.0
